+++ server/src/controllers/api/study.controller.ts
@@ -16,7 +16,7 @@ import { StudyCreation } from 'shared/dtos/study.dto';
 import { UserStudy } from '../../schemas/userstudy.schema';
 import { UserStudyService } from '../../services/userstudy.service';
 import { UserService } from '../../services/user.service';
-import { Roles } from '../../decorators/roles.decorator';
+import { Auth } from '../../guards/auth.guard';
 
 @Controller('/api/study')
 export class StudyController {
@@ -31,7 +31,7 @@ export class StudyController {
    * Get all of the studies
    */
   @Get('/')
-  @Roles('admin')
+  @Auth('admin')
   async getStudies(): Promise<Study[]> {
     return this.studyService.getStudies();
   }
@@ -40,7 +40,7 @@ export class StudyController {
    * Deterine if a study with the given name exists
    */
   @Get('/exists')
-  @Roles('admin')
+  @Auth('admin')
   async studyExists(@Query('studyName') studyName: string): Promise<boolean> {
     return this.studyService.exists(studyName);
   }
@@ -50,7 +50,7 @@ export class StudyController {
    * study.
    */
   @Get('/users')
-  @Roles('admin')
+  @Auth('admin')
   async getUserStudies(
     @Query('studyID') studyID: string,
   ): Promise<UserStudy[]> {
@@ -94,7 +94,7 @@ export class StudyController {
    * Change if the given user will have tagging access on the given study.
    */
   @Put('/user/enable')
-  @Roles('admin')
+  @Auth('admin')
   async changeUserAccess(
     @Body()
     changeRequest: {
@@ -131,7 +131,7 @@ export class StudyController {
    * newly created study.
    */
   @Post('/create')
-  @Roles('admin')
+  @Auth('admin')
   async createStudy(@Body() studyCreation: StudyCreation): Promise<Study> {
     // Make sure the study name is unique
     const exists = await this.studyService.exists(studyCreation.study.name);